% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction_functions.R
\name{extract_daily}
\alias{extract_daily}
\title{Extract daily SST values}
\usage{
extract_daily(pathin)
}
\arguments{
\item{pathin}{a character file path to the working directory (see details).}
}
\value{
a csv format file containing daily SST data for all points in the
    extraction shape file.
}
\description{
\code{extract_daily} takes a file path to img format CRW data and previosly
   created extraction shape file, extracts daily SST values and outputs to
   csv.
}
\details{
Takes a file path to img format files (the working directory where
    `img_data/` is found), previously created by (\code{\link{nc_to_img}}),
    then using the previously created extraction shape file
    (\code{\link{create_vectors}}), extracts daily SST values and writes them
    to csv.
}
\examples{
\dontrun{
extract_daily(pathin = ".")
}

}
\author{
Bart Huntley, \email{bart.huntley@dbca.wa.gov.au}

For more details see \url{https://dbca-wa.github.io/dbcaDHW/index.html}
{the dbcaDHW website}
}
